requirements.txt
Django == 4.1
psycopg2==2.9.3
django-createsuperuser==2020.12.3
django-reset-migrations==0.4.0
SQLAlchemy==1.4.40
redis==4.3.4
requests==2.27.1
celery==5.2.7    
djangorestframework==3.14.
python-dotenv==1.0.0
channels==3.0.4
daphne==3.0.2
uwsgi==2.0.22
asgiref==3.5.2
channels-redis==3.2.0
